+++ stm32/libraries/AdafruitCrypto/adafruit_base64.cpp
@@ -44,7 +44,7 @@
  * @param size      Output buffer size
  * @return  Number of bytes converted, 0 if failed
  */
-uint16_t AdafruitBase64::encode(const uint8_t* input, uint16_t inputlen, char* output, uint16_t bufsize)
+uint32_t AdafruitBase64::encode(const uint8_t* input, uint32_t inputlen, char* output, uint32_t bufsize)
 {
   return Feather.sdep(SDEP_CMD_BASE64_ENCODE, inputlen, input, &bufsize, output) ? bufsize : 0;
 }
@@ -57,7 +57,7 @@ uint16_t AdafruitBase64::encode(const uint8_t* input, uint16_t inputlen, char* o
 // * @param size      Output buffer size
 // * @return  Number of bytes converted, 0 if failed
 // */
-//uint16_t AdafruitBase64::base64Decode(const char* input, uint16_t inputlen, uint8_t* output, uint16_t size)
+//uint32_t AdafruitBase64::base64Decode(const char* input, uint32_t inputlen, uint8_t* output, uint32_t size)
 //{
 //  return Feather.sdep(SDEP_CMD_BASE64_DECODE, inputlen, input, &size, output) ? size : 0;
 //}